backend/requirements.txt
a2a-sdk>=0.3.0
fastapi>=0.115.2
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
python-jose[cryptography]>=3.3.0
httpx>=0.28.1
python-multipart>=0.0.12
